backend/requirements.txt
flask>=2.3.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
pandas>=2.0.0
openpyxl>=3.1.0
